riscv: dts: thead: Add TH1520 pin control nodes
Add nodes for pin controllers on the T-Head TH1520 RISC-V SoC.

Unfortunately there is not yet any clock driver for the TH1520 so clocks
are modeled by dummy fixed clocks in the device tree. The apb_clk is already
there for two of the pin controllers, but we need to add the missing
aonsys_clk for the always-on pin controller.

Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>