ti: vpe: Add the VIP driver
VIP stands for Video Input Port, it can be found on devices such as
DRA7xx and provides a parallel interface to a video source such as
a sensor or TV decoder.  Each VIP can support two inputs (slices) and
a SoC can be configured with a variable number of VIP's.
Each slice can supports two ports each connected to its own
sub-device.

[Bartosz:
  - VFL_TYPE_GRABBER was removed so replace it with VFL_TYPE_VIDEO,
  - checkpatch fixes]

Signed-off-by: Benoit Parrot <bparrot@ti.com>
[Tomi: fixes to the driver code]
Signed-off-by: Tomi Valkeinen <tomba@kernel.org>
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
[forward port vip driver from ti-linux-5.10.y to ti-linux-6.1.y]
Signed-off-by: Drew Fustini <dfustini@baylibre.com>
Link: https://www.spinics.net/lists/devicetree/msg355764.html